requirements.txt
Flask==0.10.1
Flask-HTTPAuth==2.2.0
Flask-RESTful==0.2.11
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.18
PyYAML==3.10
SQLAlchemy==0.9.3
Werkzeug==0.9.4
aniso8601==0.82
itsdangerous==0.23
pyaml==13.12.0
python-mimeparse==0.1.4
pytz==2013.9
pyzmq==14.0.1
requests==2.2.1
six==1.5.2
wsgiref==0.1.2
